data_science/requirements.txt
jupyter
matplotlib
geopandas
shapely
folium
azure-storage-blob
vincenty